fix(gpuagent): widen ECC block iterator to uint64_t for amdsmi 27.1.0 (#97)
amdsmi PR rocm-systems#10973 expanded AMDSMI_GPU_BLOCK_LAST from bit 18
(MPIO) to bit 38 (UCIE_PCS). The uint32_t ECC loop iterator overflows
at bit 31 and wraps to 0, creating an infinite loop (same bug RDC hit
and fixed per rocm-systems#11722 changelog). Widen to uint64_t in all 4
sites. Update vendored amdsmi libs+header to 27.1.0+4c2f7505.

+++ sw/nic/gpuagent/api/smi/amdsmi/smi_state.cc
@@ -185,7 +185,7 @@ smi_state::smi_watcher_update_all_watch_fields_(uint32_t gpu_id,
     }
 
     // get correctable and uncorrectable total error count beforehand
-    for (uint32_t b = AMDSMI_GPU_BLOCK_FIRST; b <= AMDSMI_GPU_BLOCK_LAST;
+    for (uint64_t b = AMDSMI_GPU_BLOCK_FIRST; b <= AMDSMI_GPU_BLOCK_LAST;
          b = b * 2) {
         // initialize ec to all 0s
         ec = { 0 };
